+++ app/src/main/java/com/debiandroid/desktop/ui/theme/Color.kt
@@ -2,27 +2,27 @@ package com.debiandroid.desktop.ui.theme
 
 import androidx.compose.ui.graphics.Color
 
-val Purple80 = Color((((((((((((((((((((0xFFD0BCFF).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt())
-val PurpleGrey80 = Color((((((((((((((((((((0xFFCCC2DC).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt())
-val Pink80 = Color((((((((((((((((((((0xFFEFB8C8).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt())
+val Purple80 = Color(0xFFD0BCFF)
+val PurpleGrey80 = Color(0xFFCCC2DC)
+val Pink80 = Color(0xFFEFB8C8)
 
-val Purple40 = Color((((((((((((((((((((0xFF6650a4).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt())
-val PurpleGrey40 = Color((((((((((((((((((((0xFF625b71).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt())
-val Pink40 = Color((((((((((((((((((((0xFF7D5260).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt())
+val Purple40 = Color(0xFF6650a4)
+val PurpleGrey40 = Color(0xFF625b71)
+val Pink40 = Color(0xFF7D5260)
 
-val SurfaceDark = Color((((((((((((((((((((0xFF1C1B1F).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt())
-val SurfaceLight = Color((((((((((((((((((((0xFFFFFBFE).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt())
+val SurfaceDark = Color(0xFF1C1B1F)
+val SurfaceLight = Color(0xFFFFFBFE)
 
-val DebianRed = Color((((((((((((((((((((0xFFA80030).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt())
-val DebianRedDark = Color((((((((((((((((((((0xFFD00040).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt())
-val DebianRedLight = Color((((((((((((((((((((0xFFFF6B8A).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt())
+val DebianRed = Color(0xFFA80030)
+val DebianRedDark = Color(0xFFD00040)
+val DebianRedLight = Color(0xFFFF6B8A)
 
-val TerminalBackground = Color((((((((((((((((((((0xFF1E1E1E).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt())
-val TerminalText = Color((((((((((((((((((((0xFFD4D4D4).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt())
-val TerminalGreen = Color((((((((((((((((((((0xFF4EC9B0).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt())
-val TerminalYellow = Color((((((((((((((((((((0xFFDCDCAA).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt())
-val TerminalBlue = Color((((((((((((((((((((0xFF569CD6).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt())
+val TerminalBackground = Color(0xFF1E1E1E)
+val TerminalText = Color(0xFFD4D4D4)
+val TerminalGreen = Color(0xFF4EC9B0)
+val TerminalYellow = Color(0xFFDCDCAA)
+val TerminalBlue = Color(0xFF569CD6)
 
-val StatusRunning = Color((((((((((((((((((((0xFF4CAF50).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt())
-val StatusStopped = Color((((((((((((((((((((0xFF9E9E9E).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt())
-val StatusError = Color((((((((((((((((((((0xFFE53935).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt()).toInt())
+val StatusRunning = Color(0xFF4CAF50)
+val StatusStopped = Color(0xFF9E9E9E)
+val StatusError = Color(0xFFE53935)
